+++ apps/kyb-app/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Updated dependencies
   - @ballerine/common@0.9.6
   - @ballerine/workflow-browser-sdk@0.6.9
+  - @ballerine/blocks@0.2.3
 
 ## 0.3.8